land swap intervention + postgres setup

+++ db/migrate/20210105100559_create_lands.rb
@@ -1,7 +1,7 @@
 class CreateLands < ActiveRecord::Migration[6.0]
   def change
     create_table :lands do |t|
-      t.decimal :area
+      t.float :area
       t.belongs_to :farm
       t.belongs_to :land_type
       t.boolean :sprayed
