+++ db/migrate/20210126140450_create_hedgerows.rb
@@ -1,7 +1,7 @@
 class CreateHedgerows < ActiveRecord::Migration[6.0]
   def change
     create_table :hedgerows do |t|
-      t.decimal :length
+      t.float :length
       t.belongs_to :farm
       t.belongs_to :hedgerow_type
       t.timestamps
